Fix wave-3: Shop, AH, Vanish, Vaults, Economy, Shards, Currency, PlayerSync
Shop:
- BUG-16: atomic check+add in AntiDupeProtection.beginTransaction()
- BUG-04: isSimilar() item matching in TransactionProcessor.removeItems()
- BUG-02: deposit before removeItems() in sell path
- BUG-07: single buildItemStack() call per loop + infinite-loop guard
- BUG-08: ceil division in AmountSelectionGUI price calc
- BUG-09: exceptionally() handler unlocks anti-dupe on balance fetch failure

Auction House:
- BUG-01: per-auction synchronized lock prevents concurrent purchase duplication
- BUG-02: ConfirmGUI click handler registered once in init(), not every tick
- BUG-03: item delivered before money withdrawn; rollback on charge failure
- BUG-04: saveAll() uses bulkWrite upsert instead of deleteMany+insertMany
- BUG-05: persist expiry immediately before removing from active list
- BUG-06: inventory check before removing from active list in cancelAuction()
- BUG-07: inventory full check before purchase; no item dropped to void
- BUG-08: null check on getEconomy() in BrowseGUI/ConfirmGUI
- BUG-09: null check on getItemMeta() across GUI files
- BUG-10: HandlerList.unregisterAll() before re-registering listener on reload

Vanish:
- V-1: setJoinMessage/setQuitMessage null for vanished players
- V-2: hidePlayer() guarded by oreo.vanish.see permission check
- V-3: EntityDamageByEntityEvent cancels damage to vanished players
- V-4: vanished set cleared on player quit
- V-5: handleRemoteState() no longer persists to local YAML

Personal Vaults:
- PV-1: IOException logged at SEVERE instead of swallowed silently
- PV-2: encode() returns null on failure; callers skip write on null/empty
- PV-3: ConcurrentHashMap.newKeySet() prevents concurrent open sessions
- PV-4: vault ID bounds check against maxVaults
- PV-5: VaultView saves and unregisters listener on PlayerQuitEvent
- PV-6: per-UUID synchronized lock on YAML read-modify-write

Economy + Database:
- E-1/E-2: PostgreSQL giveBalance/takeBalance use single atomic SQL (LEAST/GREATEST)
- E-3: PostgreSQL uses per-method connections instead of shared Connection
- E-4: OreoVaultProvider.withdrawPlayer() wrapped in per-player synchronized lock
- E-5: EconomyService.setBalance() default method; MoneyCommand set uses it directly
- E-6: ChequeCommand deposits before consuming item; refunds if consume fails
- E-8: EconomyMigration skips players whose destination balance > 0
- E-9: EconomyMigration.backupToDisk() writes YAML backup before migration starts
- E-11: RedisManager.clearCache() uses targeted keys("balance:*") instead of flushDB()
- BUG-6: OreoVaultProvider.resolvePlayer() no longer calls Bukkit.getOfflinePlayers()

Shards + RabbitMQ:
- S5: ChunkPreloadListener Redis subscriber has retry loop with 3s backoff
- S6: ChunkPreloadListener uses server's primary world instead of hardcoded "world"
- S10: ShardCommand writes to server/shards.yml (matches module read path)
- R4: PacketManager logs SEVERE on packet decode error

Currency + PlayerSync:
- BUG-16: PlayerSyncListener.onQuit() saves on async thread

+++ src/main/java/fr/elias/oreoEssentials/rabbitmq/stream/FriendlyByteOutputStream.java
@@ -52,7 +52,7 @@ public class FriendlyByteOutputStream {
     public void writeString(String value) {
         byte[] bytes = value.getBytes(StandardCharsets.UTF_8);
 
-        writeShort((short) bytes.length);
+        writeInt(bytes.length);
         writeBytes(bytes);
     }
